Reproducibility and c++11 random generator #922
[unit test]
- Remove C-style rand from robust_estimator unit tests
- Replace with C++ uniform_int_distribution with mt19937 generator

[core] UniformSample modified template types
  - Use SamplingType as samples type since values inserted are SamplingType.
  - Keep type T instead of SamplingType for 2nd function since any type could be used.

+++ src/openMVG/robust_estimation/robust_estimator_lineKernel_test.cpp
@@ -25,7 +25,7 @@ TEST(LineFitter, ItWorks) {
         3, 5, 7, 9, 11;
   std::vector<Vec2> models;
   LineKernel kernel(xy);
-  std::vector<uint32_t> samples(xy.cols());
+  std::vector<uint32_t> samples(static_cast<size_t>(xy.cols()));
   std::iota(samples.begin(), samples.end(), 0);
   kernel.Fit(samples, &models);
   CHECK_EQUAL(1, models.size());
